Add plot for unconstrained mode
