+++ PopPUNK/utils.py
@@ -445,3 +445,48 @@ def createOverallLineage(rank_list, lineage_clusters):
         overall_lineages['overall'][isolate] = overall_lineage
 
     return overall_lineages
+
+
+def transformLine(s, mean0, mean1):
+    """Return x and y co-ordinates for traversing along a line between mean0 and mean1, parameterised by
+    a single scalar distance s from the start point mean0.
+
+    Args:
+        s (float)
+            Distance along line from mean0
+        mean0 (numpy.array)
+            Start position of line (x0, y0)
+        mean1 (numpy.array)
+            End position of line (x1, y1)
+    Returns:
+        x (float)
+            The Cartesian x-coordinate
+        y (float)
+            The Cartesian y-coordinate
+    """
+    tan_theta = (mean1[1] - mean0[1]) / (mean1[0] - mean0[0])
+    x = mean0[0] + s * (1/np.sqrt(1+tan_theta))
+    y = mean0[1] + s * (tan_theta/np.sqrt(1+tan_theta))
+
+    return np.array([x, y])
+
+
+def decisionBoundary(intercept, gradient):
+    """Returns the co-ordinates where the triangle the decision boundary forms
+    meets the x- and y-axes.
+
+    Args:
+        intercept (numpy.array)
+            Cartesian co-ordinates of point along line (:func:`~transformLine`)
+            which intercepts the boundary
+        gradient (float)
+            Gradient of the line
+    Returns:
+        x (float)
+            The x-axis intercept
+        y (float)
+            The y-axis intercept
+    """
+    x = intercept[0] + intercept[1] * gradient
+    y = intercept[1] + intercept[0] / gradient
+    return(x, y)
